Move all the guts of extension.js into app.js.
extension.js now just loads modules using GJS's import system and reexports them
for app.js's consumption through a System.register() style module.
